[Fix] Organize code
