+++ src/components/Pool/LoadingList/LoadingList.types.ts
@@ -1,0 +1,3 @@
+export interface LoadingListProps {
+  itemsAmount?: number
+}
