+++ src/components/Pool/SortByDropdown/SortByDropdown.styles.ts
@@ -1,0 +1,23 @@
+import styled from 'styled-components'
+import { Text, Flex } from 'rebass/styled-components'
+
+export const List = styled.ul`
+  margin: 0;
+  padding: 0;
+`
+
+export const ListItem = styled.li`
+  cursor: pointer;
+  & + & {
+    margin-top: 28px;
+  }
+`
+
+export const StyledText = styled(Text)`
+  text-transform: uppercase;
+`
+
+export const StyledFlex = styled(Flex)`
+  cursor: pointer;
+  gap: 4px;
+`
